Fix ParallelGroup
Altering the list while iterating, without incrementing the erased member, and the list being altered by other RunNext calls caused memory access issues.
This fix does not affect the performance when there are only module members in ParallelGroup, so the evaluations and the combined_test/test1 results don't need to be updated.